Added: Search data by id

+++ src/main/java/com/greetingapplication/repository/Repo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Repo extends JpaRepository <GreetEntity, Long> {
+public interface Repo extends JpaRepository <GreetEntity, Integer> {
 
 }
